Implement signal decorator.
context: aossi.deco

aossi.deco.signal is a decorator that wraps a decoratable callable
in a DecoSignal instance, which in turn contains decorator
methods for further attaching other methods/functions.

These DecoSignal decorator methods only do connection and do not
in any way modify the original callable.

Unit tests: aossi.tests.Test_deco

Other changes
-------------
Added support for 'weak' keyword to aossi.signal.Signal class. If
this is set to False, uses a hard reference to the passed in
callable.

Added support for 'weakcondf' keyword to aossi.signal.Signal class. If
this is set to False, uses a hard reference for all conditional functions
sent as part of the 'choose' keyword argument.