Implement object subscription registration.
context: aossi.subscription

Based on the idea that any object can be used as an 'issue'
identifier, can register any callable (as defined by
aossi.misc.iscallable) with a subscription issue.

The subscription system is heavily based on a similar system
found in the dojo ajax toolkit/framework v0.3.1, though of course,
being a mere frontend to the basic aossi signal slot implementation, 
it is much more flexible.