Implement 'cascade' policy in Signal class.
The chooser must return either None or a list of
functions to execute. The default chooser function --
aossi.misc.ChooseCallable -- accepts a policy of
'cascade' which returns a sequence of all functions
that have their choice functions evaluate to True. If
any choice functions raises a StopCascade exception,
any following choice functions will not be evaluated.

Note that only the return value of the last callable
in the returned sequence is returned.