[aossi @ 2fa138e84bc143028b865ade28a90234f8b4395f] Implement 'cascade' policy in Signal class.
The chooser must return either None or a list of
functions to execute. The default chooser function --
aossi.misc.ChooseCallable -- accepts a policy of
'cascade' which returns a sequence of all functions
that have their choice functions evaluate to True.

Note that only the return value of the last callable
in the returned sequence is returned.

Original author: "Ariel De Ocampo <arieldeocampo@gmail.com>"
Date: 2006-06-26 15:36:36-04:00