[aossi @ 145296f11e8f59f7c1450aa7913426f407430ec3] Implement 'cascade' policy in Signal class.
The chooser must return either None or a list of
functions to execute. The default chooser function --
aossi.misc.ChooseCallable -- accepts a policy of
'cascade' which returns a sequence of all functions
that have their choice functions evaluate to True. If
any choice functions raises a StopCascade exception,
any following choice functions will not be evaluated.

Note that only the return value of the last callable
in the returned sequence is returned.

Original author: "Ariel De Ocampo <arieldeocampo@gmail.com>"
Date: 2006-07-19 14:13:48-04:00