Added support for calling method names instead of methods directly.
Normally, assuming a method signal, connected methods are stored
and called directly. Setting 'callmethod' on either the signal
or on a per-slot basis stores a function that retrieves the
method off of the instance by name and calls that. In effect, this
allows for slot methods to be used in subclasses without having
to redo or mess around with a super class' signals.