BUGFIX: streamin connection does not know to handle methods.
When used via decorator or when constructing a class, any
methods are actually functions when passed to the signal.
Thus, there is no way to determine if a function being
associated with a signal is meant to be a function or not.
The streamin connection facility thus is changed to only add
the keyword arguments as an anonymous argument instead of
transforming the entire args tuple into a list.