BUGFIX: streamin connection does not handle methods correctly.
Will now look at the first argument in the argument list and decide if the
signal is a method or not. This is accomplished by retrieving an attribute
from the first argument of the argument list with the same name as the
signal and returning the signal placed on it. If there are no arguments, or
if the first argument of the argument list is not a signal, then the main
signal is assumed not to be a method.