First implementation of base abstract class EqObj.
This is a very simple class that provides mechanisms to compare or
validate a python object.

It supports three ways to perform a comparison:

1. bool casting of an EqObj object
   - If no object was specified as a base object when instantiating
     an EqObj object, the None object is used by default
2. Calling an EqObj optionally passing an object to compare/validate
3. Directly using the == operator

A mixin or subclass would typically override __transform__ or __compare__
methods. When doing a comparison/validation, the __eq__ method will pass
its argument to transform whose return value is then passed to __compare__.
The return value of the __compare__ method is what is returned.