Override bitwise and/or operators to construct boolean EqObj objects.
Context: eqobj.core

New Classes
=============

BooleanOperation
-----------------
Base class for boolean EqObj objects. Enforces initialization with
two objects. The only requirement for these two objects are that
they are callable.

OrObj
-----
Performs a boolean OR operation on the results of the two stored
callables.

AndObj
-------
The same as OrObj but performs a boolean AND operation instead.