Move streamin slot from after category to around category.
It doesn't make sense to change the input arguments to the signal after the
signal has already been invoked and a return value created. So the streamin
slot is moved to the around category to enable manipulation of argument
list, values. It cannot be placed in the before category as those callables
cannot modify the arguments that get passed to the signal.